Fix incorrectly named "RTC" to "usec timer".
ipodhw was incorrectly referring to the usec timer as the RTC,
when the RTC also exists and is a separate register.

This was simply a naming error, the usec timer was always being used as
the usec timer, just mis-named as the RTC.

This fixes the #defines, as well as the name in the ipod_t struct.